+++ app/page.jsx
@@ -1,6 +1,23 @@
 "use client"
-import { NavBar } from "@/components/navBar"
 import { useState } from "react";
+import { useTheme } from "next-themes";
+import Link from "next/link";
+import { BoxIcon, MoonIcon, SunIcon } from "@radix-ui/react-icons";
+import { SketchPicker, ChromePicker } from "react-color";
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog"
+import { Badge } from "@/components/ui/badge";
+import { NavBar } from "@/components/navBar";
 import { RotateIcons } from "@/components/rotateIcons";
 import {
   DropdownMenu,
@@ -8,11 +25,7 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { useTheme } from "next-themes";
-import { MoonIcon, SunIcon } from "@radix-ui/react-icons";
 import { Button } from "@/components/ui/button";
-import { SketchPicker, ChromePicker } from "react-color";
-import Link from "next/link";
 
 export default function Home() {
   const { setTheme } = useTheme();
@@ -46,32 +59,49 @@ export default function Home() {
     setTheme(color === "dark" ? "dark" : "light");
   };
 
-
-  const handleTextColorChange = (color) => {
-    setTextColor(color.hex);
+  const handleColorChange = (color, type) => {
+    if (type === "text") {
+      setTextColor(color.hex);
+    } else if (type === "bg") {
+      setBgColor(color.hex);
+    } else if (type === "primary") {
+      setPrimaryColor(color.hex);
+    } else if (type === "secondary") {
+      setSecondaryColor(color.hex);
+    } else if (type === "accent") {
+      setAccentColor(color.hex);
+    }
   };
 
-  const handleBgColorChange = (color) => {
-    setBgColor(color.hex);
+  const handleCloseColorPicker = () => {
+    setSelectedColor(null);
   };
 
-  const handlePrimaryColorChange = (color) => {
-    setPrimaryColor(color.hex);
-  };
-
-  const handleSecondaryColorChange = (color) => {
-    setSecondaryColor(color.hex);
-  };
-
-  const handleAccentColorChange = (color) => {
-    setAccentColor(color.hex);
-  };
   return (
     <main>
       <NavBar />
-      <div>
+      <div className="mt-32">
         <div className="md:text-left text-center m-10">
-          <h1 className="md:text-7xl text-5xl font-bold font-poppins mt-32 md:mt-20 md:mr-44 md:ml-20">Visualize your <span className="font-borel" style={{ color: accentColor }}>design</span> on the <span className="font-effect">real</span> web.</h1>
+          <h1 className="md:text-7xl text-5xl font-bold font-poppins mt-32 md:mt-20 md:mr-44 md:ml-20">
+            Visualize your <span className="font-borel" style={{ color: accentColor }}>design</span> on the <span className="font-effect">real</span> web.
+          </h1>
+          <div className="md:ml-20 mt-6">
+            <AlertDialog>
+              <AlertDialogTrigger><Badge className="ml-1 px-4 py-1 font-medium" variant="secondary">✨ New update released</Badge></AlertDialogTrigger>
+              <AlertDialogContent>
+                <AlertDialogHeader>
+                  <AlertDialogTitle>✨ New update! v0.2.7</AlertDialogTitle>
+                  <AlertDialogDescription>
+                    This update focuses on polishing the UI and making it perfect before moving on to major additions like the templates page. All the changes and features of this update will be listed down in the changelog for you to check out!
+                  </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                  <AlertDialogCancel>Cancel</AlertDialogCancel>
+                  <Link target="_blank" href={"https://github.com/git-Abdul/huecraft/blob/master/Changelog.md"} className="flex justify-center md:block"><AlertDialogAction>Changelog</AlertDialogAction></Link>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
+          </div>
           <div className="mt-10 md:hidden lg:hidden block">
             <div className="flex justify-center items-center gap-3">
               <p className="md:mr-44 md:ml-20 text-xl text-zinc-400">Rotate your phone</p>
@@ -84,17 +114,18 @@ export default function Home() {
             <button style={{ backgroundColor: secondaryColor }} className="font-medium dark:text-white text-white hover:-translate-y-1 hover:scale-110 rounded-lg py-3 px-4 transition ease-in-out duration-200">Get started</button>
           </Link>
           <Link href={"/docs"}>
-            <button className="border font-medium dark:border-zinc-700 dark:hover:border-zinc-800 border-slate-500 hover:boder-slate-600 dark:text-white text-black hover:-translate-y-1 hover:scale-110 rounded-lg py-3 px-4 transition ease-in-out duration-200">Documentation</button>
+            <button className="border font-medium dark:border-zinc-700 dark:hover:border-zinc-800 border-slate-500 hover:border-slate-600 dark:text-white text-black hover:-translate-y-1 hover:scale-110 rounded-lg py-3 px-4 transition ease-in-out duration-200">Documentation</button>
           </Link>
         </div>
       </div>
 
       <div style={{ zIndex: 0 }} className="mt-20">
+        {/* Color boxes */}
         <div className="flex justify-center gap-2 mt-16 px-8">
-          <div className="rounded-lg w-48 md:h-44 h-20" style={{ backgroundColor: primaryColor }}></div>
-          <div className="rounded-lg w-48 md:h-44 h-20" style={{ backgroundColor: secondaryColor }}></div>
-          <div className="rounded-lg w-40 md:h-44 h-20" style={{ backgroundColor: bgColor }}></div>
-          <div className="rounded-lg w-36 md:h-44 h-20" style={{ backgroundColor: accentColor }}></div>
+          <div className="rounded-lg w-56 md:h-44 h-20" style={{ backgroundColor: primaryColor }}></div>
+          <div className="rounded-lg w-52 md:h-44 h-20" style={{ backgroundColor: secondaryColor }}></div>
+          <div className="rounded-lg w-44 md:h-44 h-20" style={{ backgroundColor: bgColor }}></div>
+          <div className="rounded-lg w-40 md:h-44 h-20" style={{ backgroundColor: accentColor }}></div>
           <div className="rounded-md w-4 md:h-44 h-20" style={{ backgroundColor: textColor }}></div>
         </div>
         <div className="flex justify-center gap-2 px-8 py-2">
@@ -106,11 +137,10 @@ export default function Home() {
         </div>
       </div>
 
-      {/*MenuBar*/}
-
+      {/* MenuBar */}
       <div className="text-center items-center justify-center flex invisible md:visible">
-        <div className="flex justify-center gap-2 item-blur bg-zinc-900 mx-32 p-5 fixed mt-20 bottom-0 mb-5 rounded-lg" style={{ zIndex: "200" }}>
-          {/* Text Button */}
+        <div className="flex justify-center gap-2 backdrop-blur-lg bg-white/10 dark:bg-zinc-800/30 bg-zinc-900 mx-32 p-5 fixed mt-20 bottom-0 mb-5 rounded-lg" style={{ zIndex: "5" }}>
+          {/* Color buttons */}
           <Button
             className="py-6 px-7 text-md dark:text-white text-black"
             style={{ backgroundColor: "#202020", color: "#fff" }}
@@ -118,8 +148,6 @@ export default function Home() {
           >
             Text
           </Button>
-
-          {/* Background Button */}
           <Button
             className="py-6 px-7 text-md dark:text-white"
             style={{ backgroundColor: bgColor }}
@@ -127,8 +155,6 @@ export default function Home() {
           >
             Background
           </Button>
-
-          {/* Primary Button */}
           <Button
             className="py-6 px-7 text-md dark:text-white"
             style={{ backgroundColor: primaryColor }}
@@ -136,8 +162,6 @@ export default function Home() {
           >
             Primary
           </Button>
-
-          {/* Secondary Button */}
           <Button
             className="py-6 px-7 text-md dark:text-white"
             style={{ backgroundColor: secondaryColor }}
@@ -145,8 +169,6 @@ export default function Home() {
           >
             Secondary
           </Button>
-
-          {/* Accent Button */}
           <Button
             className="py-6 px-7 text-md dark:text-white"
             style={{ backgroundColor: accentColor }}
@@ -155,12 +177,11 @@ export default function Home() {
             Accent
           </Button>
 
+          {/* Theme dropdown */}
           <hr className="w-0.5 rounded-lg h-12 bg-gray-400" />
-
-          {/* ... other buttons */}
           <DropdownMenu>
             <DropdownMenuTrigger asChild>
-              <Button variant="secondary" size="icon" className="w-[3rem] h-[3rem]">
+              <Button variant="link" size="icon" className="w-[3rem] h-[3rem]">
                 <SunIcon className="absolute h-[2rem] w-[2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
                 <MoonIcon className="absolute h-[2rem] w-[2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
                 <span className="sr-only">Toggle theme</span>
@@ -174,30 +195,26 @@ export default function Home() {
           </DropdownMenu>
         </div>
       </div>
+
       {selectedColor && (
-        <div className="p-3 rounded-lg dark:bg-zinc-800 bg-slate-200" style={{ position: "absolute", top: "100px", left: "50%", transform: "translateX(-50%)", position: "fixed", top: "calc(50% - 150px)", left: "50%", transform: "translateX(-50%)", zIndex: "200", }} >
+        <div className="p-3 rounded-lg backdrop-blur-lg bg-white/10 dark:bg-zinc-800/30" style={{ position: "absolute", top: "100px", left: "50%", transform: "translateX(-50%)", position: "fixed", top: "calc(50% - 150px)", left: "50%", transform: "translateX(-50%)", zIndex: "200" }}>
           <ChromePicker
-            color={selectedColor === "text" ? textColor : selectedColor === "bg" ? bgColor : selectedColor === "primary" ? primaryColor : selectedColor === "secondary" ? secondaryColor : accentColor}
-            onChange={color =>
-              setSelectedColor(prevColor => {
-                if (prevColor === "text") {
-                  handleTextColorChange(color);
-                } else if (prevColor === "bg") {
-                  handleBgColorChange(color);
-                } else if (prevColor === "primary") {
-                  handlePrimaryColorChange(color);
-                } else if (prevColor === "secondary") {
-                  handleSecondaryColorChange(color);
-                } else if (prevColor === "accent") {
-                  handleAccentColorChange(color);
-                }
-                return prevColor;
-              })
+            color={
+              selectedColor === "text"
+                ? textColor
+                : selectedColor === "bg"
+                  ? bgColor
+                  : selectedColor === "primary"
+                    ? primaryColor
+                    : selectedColor === "secondary"
+                      ? secondaryColor
+                      : accentColor
             }
+            onChange={(color) => handleColorChange(color, selectedColor)}
           />
-          <Button onClick={() => setSelectedColor(null)} className="mt-2">Close</Button>
+          <Button variant="secondary" onClick={handleCloseColorPicker} className="mt-2">Close</Button>
         </div>
       )}
     </main>
-  )
+  );
 }