+++ components/DocsList.js
@@ -3,11 +3,10 @@ import { ScrollArea } from "@/components/ui/scroll-area"
 import { Badge } from "@/components/ui/badge"
 import Link from "next/link"
 
-export function DocScroll() {
+export function DocsList() {
     return (
-        <ScrollArea className="w-48 fixed h-screen text-center text-lg border-r dark:border-zinc-700 border-gray-300">
+        <ScrollArea className="w-48 fixed h-screen border-b text-center text-lg dark:border-zinc-700 border-gray-300">
             <div className="p-4 text-sm">
-                <h4 className="text-xl p-2 font-inter font-bold leading-none text-center md:block hidden">Documentation</h4>
                 <Badge variant="destructive">In Development</Badge>
                 <Link href={"/docs#introduction"}>
                 <div className="mt-6 dark:hover:bg-zinc-800 dark:bg-zinc-900 bg-slate-200 hover:bg-slate-300 transition ease-in-out duration-200 p-1 rounded-lg">